requirements.txt
fastapi
python-multipart
uvicorn
torch
torchvision
pillow